add display support
Select type of display to use. This option is a replacement for the old style -sdl/-curses/... options. Valid values for type are

sdl
Display video output via SDL (usually in a separate graphics window; see the SDL documentation for other possibilities).

curses
Display video output via curses. For graphics device models which support a text mode, QEMU can display this output using a curses/ncurses interface. Nothing is displayed when the graphics device is in graphical mode or if the graphics device does not support a text mode. Generally only the VGA device models support text mode.

none
Do not display video output. The guest will still see an emulated graphics card, but its output will not be displayed to the QEMU user. This option differs from the -nographic option in that it only affects what is done with video output; -nographic also changes the destination of the serial and parallel port data.

gtk
Display video output in a GTK window. This interface provides drop-down menus and other UI elements to configure and control the VM during runtime.

vnc
Start a VNC server on display <arg>